requirements.txt
Flask==3.1.0
mss==10.0.0
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
